+++ prowl/reddit/reddit.py
@@ -5,8 +5,6 @@ import os
 from praw import Reddit
 from dotenv import load_dotenv
 
-from prowl.utility.file import read_file_by_line
-
 load_dotenv()
 
 USER_AGENT = os.getenv('USER_AGENT')
@@ -34,13 +32,3 @@ def stream(subreddit: str) -> Reddit.submission:
 
     return REDDIT.subreddit(subreddit).stream.submissions(skip_existing=True)
 
-
-def get_subreddit(filename: str) -> str:
-    """Will collect list of subreddits from a file into a string
-
-    :param filename: the path to the file with a list of
-    :return: a `+` separated string with all subreddits
-    """
-
-    subreddit: list = read_file_by_line(filename)
-    return '+'.join(subreddit)
